ai/requirements.txt
# Core FastAPI and web framework
fastapi==0.116.1
uvicorn[standard]==0.35.0
starlette==0.47.2

# Database
sqlalchemy==2.0.41
psycopg2-binary==2.9.9
pgvector==0.4.1
pypika==0.48.9

# Pydantic for data validation
pydantic==2.11.7
pydantic-settings==2.10.1

# LLM and AI libraries
langchain==0.3.26
langchain-core==0.3.71
langchain-text-splitters==0.3.8
langchain-community==0.3.27
langsmith==0.4.8
openai==1.97.1

# Vector search and embeddings
faiss-cpu==1.7.4
sentence-transformers==5.0.0

# ML libraries
torch==2.7.1
transformers==4.53.3
tokenizers==0.21.2
onnxruntime==1.22.1
numpy==1.26.4
scikit-learn==1.7.1

# HTTP and utilities
requests==2.32.4
requests-toolbelt==1.0.0
httpx==0.28.1
aiohttp==3.12.14

# Data processing
pandas==2.3.1
datasets==2.19.2

# Configuration and environment
python-dotenv==1.1.1
python-multipart==0.0.20

# Logging and utilities
coloredlogs==15.0.1
humanfriendly==10.0
rich==13.9.4

# Other utilities
jiter==0.10.0
jsonpatch==1.33
zstandard==0.23.0
greenlet==3.2.3
typer==0.16.0
tqdm==4.67.1
